non-zero adjoint now.

+++ main_globvar.f90
@@ -11,9 +11,9 @@ CHARACTER(len=15), ALLOCATABLE :: noms(:)
 ! Mixing Ratio Vectors 
 REAL, ALLOCATABLE :: pq_c(:,:)
 ! Tangent Linear/Adjoint Matrix
-REAL, ALLOCATABLE :: TLM(:,:,:), ADJ(:,:,:)
+REAL*8, ALLOCATABLE :: TLM(:,:,:), ADJ(:,:,:)
 ! Sensitivity Vector 
-REAL, ALLOCATABLE :: hatJ(:,:)
+REAL*8, ALLOCATABLE :: hatJ(:,:)
 
 END MODULE main_globvar
